requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=10.0.0
python-decouple>=3.8
whitenoise>=6.5.0
gunicorn>=21.2.0
